Add tests for prize edit and remove
