+++ src/commands/manage/prize/edit.test.ts
@@ -6,11 +6,11 @@ import Prize from '../../../entity/prize';
 import Server from '../../../entity/server';
 import ShamedMember from '../../../entity/shamed-member';
 import User from '../../../entity/user';
-import { createServer, createClient, createMessage, createChannel } from '../../../../test/test-helpers';
-import { addPrize, Arguments } from './add';
+import { createServer, createClient, createMessage, createChannel, createPrize } from '../../../../test/test-helpers';
+import { editPrize, Arguments } from './edit';
 import { EMOJI_ERROR, EMOJI_INCORRECT_PERMISSIONS, EMOJI_JOB_WELL_DONE } from '../../../utils/emoji';
 
-describe('commands/manage/prize/add', () => {
+describe('commands/manage/prize/edit', () => {
   beforeEach(async done => {
     await createConnection({
       type: 'sqlite',
@@ -33,10 +33,12 @@ describe('commands/manage/prize/add', () => {
 
   it(`should require permissions`, async done => {
     const server = await createServer();
+    const prize = await createPrize(server);
 
     const args: Arguments = {
       client: createClient(),
       message: await createMessage({ server }),
+      id: prize.id,
       description: 'A hellhound',
       reactionEmoji: '🐺',
       price: 3,
@@ -46,7 +48,7 @@ describe('commands/manage/prize/add', () => {
       reactions: {},
     };
 
-    const response = await addPrize(args);
+    const response = await editPrize(args);
     expect(response).toBe(`${EMOJI_INCORRECT_PERMISSIONS} You ain't got permission to do that!`);
 
     done();
@@ -54,10 +56,12 @@ describe('commands/manage/prize/add', () => {
 
   it('should require redeem-channel being set', async done => {
     const server = await createServer();
+    const prize = await createPrize(server);
 
     const args: Arguments = {
       client: createClient(),
       message: await createMessage({ server, permission: 'ADMINISTRATOR' }),
+      id: prize.id,
       description: 'A hellhound',
       reactionEmoji: '🐺',
       price: 3,
@@ -67,7 +71,7 @@ describe('commands/manage/prize/add', () => {
       reactions: {},
     };
 
-    const response = await addPrize(args);
+    const response = await editPrize(args);
     expect(response).toBe(`${EMOJI_ERROR} You need to set the \`redeem-channel\` config before using prizes.`);
 
     done();
@@ -76,6 +80,7 @@ describe('commands/manage/prize/add', () => {
   it('should require description', async done => {
     const server = await createServer();
     const channel = createChannel('redeem');
+    const prize = await createPrize(server);
 
     server.config.redeemChannelId = channel.id;
     await server.config.save();
@@ -83,6 +88,7 @@ describe('commands/manage/prize/add', () => {
     const args: Arguments = {
       client: createClient(),
       message: await createMessage({ server, serverChannels: [channel], permission: 'ADMINISTRATOR' }),
+      id: prize.id,
       description: '',
       reactionEmoji: '🐺',
       price: 3,
@@ -92,7 +98,7 @@ describe('commands/manage/prize/add', () => {
       reactions: {},
     };
 
-    const response = await addPrize(args);
+    const response = await editPrize(args);
     expect(response).toBe(`${EMOJI_ERROR} Description required!`);
 
     done();
@@ -101,6 +107,7 @@ describe('commands/manage/prize/add', () => {
   it('should require reaction emoji', async done => {
     const server = await createServer();
     const channel = createChannel('redeem');
+    const prize = await createPrize(server);
 
     server.config.redeemChannelId = channel.id;
     await server.config.save();
@@ -108,6 +115,7 @@ describe('commands/manage/prize/add', () => {
     const args: Arguments = {
       client: createClient(),
       message: await createMessage({ server, serverChannels: [channel], permission: 'ADMINISTRATOR' }),
+      id: prize.id,
       description: 'A hellhound',
       reactionEmoji: '',
       price: 3,
@@ -117,7 +125,7 @@ describe('commands/manage/prize/add', () => {
       reactions: {},
     };
 
-    const response = await addPrize(args);
+    const response = await editPrize(args);
     expect(response).toBe(`${EMOJI_ERROR} Reaction emoji required!`);
 
     done();
@@ -126,6 +134,7 @@ describe('commands/manage/prize/add', () => {
   it('should require price of 1 or more', async done => {
     const server = await createServer();
     const channel = createChannel('redeem');
+    const prize = await createPrize(server);
 
     server.config.redeemChannelId = channel.id;
     await server.config.save();
@@ -133,6 +142,7 @@ describe('commands/manage/prize/add', () => {
     const args: Arguments = {
       client: createClient(),
       message: await createMessage({ server, serverChannels: [channel], permission: 'ADMINISTRATOR' }),
+      id: prize.id,
       description: 'A hellhound',
       reactionEmoji: '🐺',
       price: 0,
@@ -142,13 +152,13 @@ describe('commands/manage/prize/add', () => {
       reactions: {},
     };
 
-    const response = await addPrize(args);
+    const response = await editPrize(args);
     expect(response).toBe(`${EMOJI_ERROR} Price must be 1 or more!`);
 
     done();
   });
 
-  it('should add the prize', async done => {
+  it('should require a valid id', async done => {
     const server = await createServer();
     const channel = createChannel('redeem');
 
@@ -158,6 +168,34 @@ describe('commands/manage/prize/add', () => {
     const args: Arguments = {
       client: createClient(),
       message: await createMessage({ server, serverChannels: [channel], permission: 'ADMINISTRATOR' }),
+      id: 7,
+      description: 'A hellhound',
+      reactionEmoji: '🐺',
+      price: 1,
+      needsFetch: false,
+      careAboutQuietMode: false,
+      promisedOutput: null,
+      reactions: {},
+    };
+
+    const response = await editPrize(args);
+    expect(response).toBe(`${EMOJI_ERROR} Couldn't find that prize, are you sure \`${args.id}\` is the right ID?`);
+
+    done();
+  });
+
+  it('should add the prize', async done => {
+    const server = await createServer();
+    const channel = createChannel('redeem');
+    const prize = await createPrize(server);
+
+    server.config.redeemChannelId = channel.id;
+    await server.config.save();
+
+    const args: Arguments = {
+      client: createClient(),
+      message: await createMessage({ server, serverChannels: [channel], permission: 'ADMINISTRATOR' }),
+      id: prize.id,
       description: 'A hellhound',
       reactionEmoji: '🐺',
       price: 10,
@@ -167,7 +205,7 @@ describe('commands/manage/prize/add', () => {
       reactions: {},
     };
 
-    const response = await addPrize(args);
+    const response = await editPrize(args);
     expect(response).toBe(`${EMOJI_JOB_WELL_DONE} Done!`);
     await server.reload();
     expect(server.prizes).toHaveLength(1);
