+++ src/commands/manage/prize/remove.test.ts
@@ -6,11 +6,11 @@ import Prize from '../../../entity/prize';
 import Server from '../../../entity/server';
 import ShamedMember from '../../../entity/shamed-member';
 import User from '../../../entity/user';
-import { createServer, createClient, createMessage, createChannel } from '../../../../test/test-helpers';
-import { addPrize, Arguments } from './add';
+import { createServer, createClient, createMessage, createChannel, createPrize } from '../../../../test/test-helpers';
+import { removePrize, Arguments } from './remove';
 import { EMOJI_ERROR, EMOJI_INCORRECT_PERMISSIONS, EMOJI_JOB_WELL_DONE } from '../../../utils/emoji';
 
-describe('commands/manage/prize/add', () => {
+describe('commands/manage/prize/remove', () => {
   beforeEach(async done => {
     await createConnection({
       type: 'sqlite',
@@ -33,20 +33,19 @@ describe('commands/manage/prize/add', () => {
 
   it(`should require permissions`, async done => {
     const server = await createServer();
+    const prize = await createPrize(server);
 
     const args: Arguments = {
       client: createClient(),
       message: await createMessage({ server }),
-      description: 'A hellhound',
-      reactionEmoji: '🐺',
-      price: 3,
+      id: prize.id,
       needsFetch: false,
       careAboutQuietMode: false,
       promisedOutput: null,
       reactions: {},
     };
 
-    const response = await addPrize(args);
+    const response = await removePrize(args);
     expect(response).toBe(`${EMOJI_INCORRECT_PERMISSIONS} You ain't got permission to do that!`);
 
     done();
@@ -54,26 +53,25 @@ describe('commands/manage/prize/add', () => {
 
   it('should require redeem-channel being set', async done => {
     const server = await createServer();
+    const prize = await createPrize(server);
 
     const args: Arguments = {
       client: createClient(),
       message: await createMessage({ server, permission: 'ADMINISTRATOR' }),
-      description: 'A hellhound',
-      reactionEmoji: '🐺',
-      price: 3,
+      id: prize.id,
       needsFetch: false,
       careAboutQuietMode: false,
       promisedOutput: null,
       reactions: {},
     };
 
-    const response = await addPrize(args);
+    const response = await removePrize(args);
     expect(response).toBe(`${EMOJI_ERROR} You need to set the \`redeem-channel\` config before using prizes.`);
 
     done();
   });
 
-  it('should require description', async done => {
+  it('should require a valid id', async done => {
     const server = await createServer();
     const channel = createChannel('redeem');
 
@@ -83,24 +81,23 @@ describe('commands/manage/prize/add', () => {
     const args: Arguments = {
       client: createClient(),
       message: await createMessage({ server, serverChannels: [channel], permission: 'ADMINISTRATOR' }),
-      description: '',
-      reactionEmoji: '🐺',
-      price: 3,
+      id: 7,
       needsFetch: false,
       careAboutQuietMode: false,
       promisedOutput: null,
       reactions: {},
     };
 
-    const response = await addPrize(args);
-    expect(response).toBe(`${EMOJI_ERROR} Description required!`);
+    const response = await removePrize(args);
+    expect(response).toBe(`${EMOJI_ERROR} Couldn't find that prize, are you sure \`${args.id}\` is the right ID?`);
 
     done();
   });
 
-  it('should require reaction emoji', async done => {
+  it('should remove the prize', async done => {
     const server = await createServer();
     const channel = createChannel('redeem');
+    const prize = await createPrize(server);
 
     server.config.redeemChannelId = channel.id;
     await server.config.save();
@@ -108,69 +105,17 @@ describe('commands/manage/prize/add', () => {
     const args: Arguments = {
       client: createClient(),
       message: await createMessage({ server, serverChannels: [channel], permission: 'ADMINISTRATOR' }),
-      description: 'A hellhound',
-      reactionEmoji: '',
-      price: 3,
+      id: prize.id,
       needsFetch: false,
       careAboutQuietMode: false,
       promisedOutput: null,
       reactions: {},
     };
 
-    const response = await addPrize(args);
-    expect(response).toBe(`${EMOJI_ERROR} Reaction emoji required!`);
-
-    done();
-  });
-
-  it('should require price of 1 or more', async done => {
-    const server = await createServer();
-    const channel = createChannel('redeem');
-
-    server.config.redeemChannelId = channel.id;
-    await server.config.save();
-
-    const args: Arguments = {
-      client: createClient(),
-      message: await createMessage({ server, serverChannels: [channel], permission: 'ADMINISTRATOR' }),
-      description: 'A hellhound',
-      reactionEmoji: '🐺',
-      price: 0,
-      needsFetch: false,
-      careAboutQuietMode: false,
-      promisedOutput: null,
-      reactions: {},
-    };
-
-    const response = await addPrize(args);
-    expect(response).toBe(`${EMOJI_ERROR} Price must be 1 or more!`);
-
-    done();
-  });
-
-  it('should add the prize', async done => {
-    const server = await createServer();
-    const channel = createChannel('redeem');
-
-    server.config.redeemChannelId = channel.id;
-    await server.config.save();
-
-    const args: Arguments = {
-      client: createClient(),
-      message: await createMessage({ server, serverChannels: [channel], permission: 'ADMINISTRATOR' }),
-      description: 'A hellhound',
-      reactionEmoji: '🐺',
-      price: 10,
-      needsFetch: false,
-      careAboutQuietMode: false,
-      promisedOutput: null,
-      reactions: {},
-    };
-
-    const response = await addPrize(args);
+    const response = await removePrize(args);
     expect(response).toBe(`${EMOJI_JOB_WELL_DONE} Done!`);
     await server.reload();
-    expect(server.prizes).toHaveLength(1);
+    expect(server.prizes).toHaveLength(0);
 
     done();
   });
